backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
langgraph
langchain-google-genai
langchain-ollama
python-multipart
pypdf
psycopg2-binary
sqlalchemy
python-dotenv
edge-tts
fpdf2
